Fixed issue with events using tiles not blocking the player.
Added "Tile Event through Fix" to scripts. It duck punches the Game_Event defined in Blizz-ABS 2 and removes the requirement that a sprite must be used by an event to have it block the player.

Blocking is now defined by
1. Tile definition. Passage of the tile used for the event.
2. Through flag. Overwrites 1. making it possible to walk over otherwise blocking events.

As noted, you can't use a non-blocking tile to create a blocking event like this.