Added script to monitor and change Deedlit's sprite.
Added "Equip Change trigger CE" script. Extends the default Game_Actor::equip. When equipment for an actor is modified, an interpreter is created and runs through a specified common event (default 5). The ID of the common event called can be modified in the module ACTORS::DEEDLIT::EQUIPMENT_CE_ID, in the "Equip Change trigger CE" script.

The Common Event added for this is 5:Update Deedlit Sprite. It checks if actor 1 (Deedlit) has any armor3 (body armor). If the actor does have any armor3 defined, Deedlit will change her sprite to Kai (default, green dress). Switch 50 (Nude) will be set to off. If the actor doesn't have an armor3 defined Deedlit will use the Kai_naked sprite, and switch 50 will be set to ON.

Since it's using an interpreter, rather than modifying a flag or variable, changes can be (and made visible) in the equipment menu.

Since it runs for all equipment changes for any/all actors, modifications should in the future be made to:
1. Identify the actor modified.
2. Make sure you wish to modify the sprite of the actor changed.
3. Make a custom call based on the actor modified.

There are also the issue with the equipment menu constantly calling the equip while the player browses the items. Since the script switches between two different items, it may be hard to detect if the change can be ignored or not, and may require modifications to the other script, rather than this one.