Made it easier to setup children through the editor.
Actor 6-9 is used to set the states at the different stages of development. Age for the stage is defined by the actors level. These should be defined before adding a child to the party.

To add new children add an extra condition to common event 11 (Add child) with the type or parent id, calling a common event that sets the stages using actors 6-9.

To add/remove states (e.g. for encumbering actor without damaging base stats) define states with + for adding a state, and - for removing one.

Added common events 13, 14, 15 for setting up the child. 13 (Human) is used as default for all children created.