Added equipment level sprite rendering.
Added the ability to guide what sprites are rendered for equipment worn.
When the game renders the sprites it sets V51 with the id of the actor being update and calls the common event 5.
As with the age progression of children, there are actors specific for each item, making it possible to modify which sprite is being used. These actors are 10-14 and needs to be modified during CE5 is executed.
To check what item is equiped, do a conditional branch with the relevant check against one of the actors (10-14) and make the relevant changes to its graphics. All 10-14 actor's grphics will be mashed together to create the final actor.
Currently Deedlit is the one capable of being modified.

Added images for the different equipment.