Added support for updating the sprite in the menu.
Previosuly any changes to the equipment wasn't visible form the menu after modifying it. It now goes through the same process as with the equipment changes in-game.

The script can be found as "draw_actor_graphic with equipment" and duckpunches Window_Base's draw_actor_graphic, making it call CE5, iterate over actors 10-14 and draw the entire thing on screen.