Added pre/post damage hooks to Common Events.
When a game actor (i.e. the player, or one of the party members), common event 55 will be called when there is a confirmed hit (but before dealing the damage) and 56 when the damage has been done on the actor.
The target will is set as actor 15, pointing to the correct actor id. Any changes to the actor hit should be done here.
The attacker will is set as the enemy troop, and changes to this battler should be done through the enemy troop commands.

Variables
30 - Target ID (Actor ID)
31 - Attacker ID (Enemy ID)
32 - Damage (HP)

Damage (Variable 32) can be modified to alter how much damaged is done to the target. Modifications to the actor and attacker are ignored.