When Deedlit gets knocked out she now turns switch 51 ON. Use this to create loss events.
Instead of having a parallel event looking at Deedlit's state, use an autorun event with the condition 51 to ON.

The logic of managing the knocked out state can be found in Common event 60.

A default knock out event has been created which takes the player to Rekka's house. Common event 61.